fixed bug, don't short ciruit transitions if one errors.
Do not return errors when applying transitions because that will short
circuit async.series so transitions will stop being applied.

Transitions are async and should return true if the document has
changed.  If a transition errors then log the error, but don't return it
because that will stop the series and the other transitions won't run.
A doc can be modified with error messages but will only get saved to the
doc if there are other (non-error) changes too.

Some time ago errors messages were added to the doc because there was no
way to surface parser errors.  Since the configuration in templates and
conditionals/expressions involves evaluating code, it's useful to
provide some clue to the user when this fails.  This isn't perfect but
it's a middleground solution.