Create transition to add info document
This new transition is responsible for adding and maintaining a sister
info document for all documents (bar info and _design documents).

So far this document contains the intial replication date and the the
latest subsequent replication date. However, in the future we will look
at moving other pieces of data here. Items that are suitable would be
things that clients never need to know about, or things that are overly
chatty and would cause unneeded replication. Obvious candidates:
 - transition metadata
 - document schema metadata (indication of schema version for example)

This commit covers the core part of the transition:
 - Loads the info doc and updates the last_updated date
 - If the doc doesn't exist try to generate it from the audit log
 - If the audit log doesn't exist generate a base info doc

Also:
 - Clarified documenation on transition API
 - Fixed a bug where a transition that exists in code but isn't
   configured doesn't get logged as disabled