New trigger to clear schedule
Limitation for now that it can only clear entire schedules, which is a
reasonable assumption for registration forms. Can be dealt with if/when
`accept_patient_reports` is fully combined with `registrations`.

This removes the need to split form validation between `registrations`
and `accept_patient_reports` sections.

This could use some tests.

Related to https://github.com/medic/medic-webapp/issues/3082