Better parsing support
We actually *do* support raw JSON, at least in theory. Other areas
may read kanso.json and think that app_settings is invalid because that
value should be a string, and break. For example, where we auto-generate
config editing.

So we also still support parsing json from a string as well, and this
will continue to be our official documentation about it for now.