Clarifying usage between two views
It is confusing. In the past "registered_patients" has been used as a
shortcut for finding patients in general. However, while patients can be
generated via a registration, they can also be created in other ways,
such as directly in the UI.

This change adds documentation around this, as well as fixing two
incorrect usages of the registration view in existing code.

medic/medic-webapp#3401