Adds init function to validate transitions
If this optional function is exported by the transition it's
called when the transition is loaded. Transition goading occurs
on start and on configuration change.

If the init function throws an error then the transition isn't
loaded, the error message is logged, and the other transitions
load as per usual.

medic/medic-webapp#3585