Store code when adding errors
When adding errors to a doc we record the message to display, but
it's helpful also to have an error code which can be inspected
programmatically without being affected by localisation and
configuration changes.

In particular this change allows us to detect when no patient is
found with the given identifier.

medic/medic-webapp#3437