Don't escape template parameters
Mustache escape HTML characters by default because usually it's
used for displaying HTML content so it's at risk for injection.
We just use it for generating SMS messages to HTML escaping isn't
relevant for us. We use angular to escaping the message content
when displayed in the webapp.

medic/medic-webapp#3795