Improve findToClear
Rework code based on conversations with Marc. See:
https://github.com/medic/medic-webapp/issues/3824#issuecomment-327068751

Specifically:
 - If there is no silence_for configured, mark all clearable tasks as
   cleared
 - If there is a silence_for configured
   - find the maximum visible date by extending the reports
     reported_date by silence_for
   - find all groups that "exist" behind that date, where exist means
     have at least one task with a due date before the date
   - and clear all clearable tasks for all those groups

We define clearable as any scheduled or pending task. Note that
previously it was just scheduled, but adding pending makes sense,
because a pending task is not any different from a scheduled one really,
it's just that gateway and friends haven't picked it up yet (maybe
gateway is down and has been down for weeks).

medic/medic-webapp#3857