Refactoring: more cleanly expose silenceReminders
The Registration transition currently calls handleReport, but only to
silenceReminders. Instead of having it call a function that also does
other things (and because of passed vars, only does that for
registrations), just call through to that one function.

This does mean there is more replicated code, but IMO it's cleaner