Fix sentinel fti integration
The `db.fti` function wasn't working due to a double slash in the
URL. It's not clear when this was broken.

Added logging so when validation functions error out the error
is logged.

Also future-proofed the `db.fti` function using the similar code
from medic-api so it works with couchdb2.

medic/medic-webapp#651